+++ src/breakpoints.js
@@ -66,7 +66,9 @@ var
 	isDisabledVerticalBreakPoints = FALSE,
 
 	LESS_APPENDIX = '-less',
-	MORE_APPENDIX = '-more';
+	MORE_APPENDIX = '-more',
+
+	_breakpointsInitialized;
 
 
 /**
@@ -301,6 +303,7 @@ $C._features.breakpoints = initBreakpoints;
 // Function declarations: ######################### ######################### ######################### ######################### ######################### ######################### #########################
 
 function initBreakpoints() {
+
 	setInterval( checkWindowOrDocumentResize, CHECK_DOCUMENT_SIZE_INTERVAL );
 
 	bind( window, 'resize', solveSizes );
@@ -308,16 +311,20 @@ function initBreakpoints() {
 	// for mobiles - on mobile devices is window size changing while scrolling content - because some panels are hiding
 	bind( document, 'touchmove', checkWindowOrDocumentResize );
 	bind( document, 'touchend', checkWindowOrDocumentResize );
+
+	_breakpointsInitialized = TRUE;
 }
 
 // on mobile devices is window size changing while scrolling content - because some panels are hiding
 function checkWindowOrDocumentResize() {
+
 	if ( getWindowWidth() != lastWinWidth || getWindowHeight() != lastWinHeight ||
 		getDocumentWidth() != lastDocWidth || getDocumentHeight() != lastDocHeight )
 		solveSizes();
 }
 
 function arrayGetIndexOfName( array, name ) {
+
 	for ( var i = 0; i < array.length; i++ )
 		if ( array[ i ].name == name )
 			return i;
@@ -367,6 +374,8 @@ function removeAllClasses( sizeAttributeName ) {
 }
 
 function solveSizes() {
+
+	if ( !_breakpointsInitialized ){ return; }
 
 	startTransactionClass();