+++ src/JS.Responsive.source.js
@@ -50,7 +50,6 @@
 	 * Constructor is PRIVATE, client must use only class methods!!!!!
 	 * @constructor
 	 * @alias JS.Responsive
-	 * @since 3.0.0
 	 *
 	 * @emit documentReady - when document becomes ready
 	 *
@@ -63,7 +62,7 @@
 	 * Library version
 	 * @const {String}
 	 */
-	$C.version = '3.0.0';
+	$C.version = '3.1.0';
 
 	// -------------------------------------------------------------------------------------------------
 	// --- CONFIG --------------------------------------------------------------------------------------